+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Statistics grouped rows: sub CQP only has one of the values [#452](https://github.com/spraakbanken/korp-frontend/issues/452)
+- Statistics CSV/TSV export broken when rows have multiple tokens [#458](https://github.com/spraakbanken/korp-frontend/issues/458)
 
 ## [9.9.0] - 2025-03-24