+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Statistics grouped rows: sub CQP only has one of the values [#452](https://github.com/spraakbanken/korp-frontend/issues/452)
 - Statistics CSV/TSV export broken when rows have multiple tokens [#458](https://github.com/spraakbanken/korp-frontend/issues/458)
+- Use attribute value without colon-suffix in stats export [#459](https://github.com/spraakbanken/korp-frontend/issues/459)
 
 ## [9.9.0] - 2025-03-24